apps/backend/pyproject.toml
[project]
name = "thelounge_bookshelf"
version = "1.0.0"
description = "ideahacks.la 2026 project"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "flask==3.1.3",
    "flask-sock>=0.7.0",
    "libusb1>=3.3.1",
    "requests>=2.33.1",
]
